routes

+++ INTEGRA-backend/routes/api/Repository/category.php
@@ -5,11 +5,11 @@ use Illuminate\Support\Facades\Route;
 
 Route::controller(CategoryController::class)->group(function () {
     Route::prefix('repository')->group(function (){
-        Route::get('categories/index', 'index');
-        Route::get('category/show/{id}', 'show');
-        Route::post('category/store', 'store');
-        Route::put('category/update/{id}', 'update');
-        Route::delete('category/destroy/{id}', 'destroy'); 
+        Route::get('categories/index', 'index')->middleware('permission:index catagory');        
+        Route::get('category/show/{id}', 'show')->middleware('permission:store catagory');
+        Route::post('category/store', 'store')->middleware('permission:show catagory');
+        Route::put('category/update/{id}', 'update')->middleware('permission:update catagory');
+        Route::delete('category/destroy/{id}', 'destroy')->middleware('permission:destroy catagory');
     });
 });
